+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fine tune the spacing of elements in menu sub items.
 - Optimize the spacing of warehouse themes on exploration pages.
 - Increase the corner angle of the pop-up window.
+- Synchronize the style of the search bar.
 
 ### 🐞 Fix